shift view generation into ViewFactory
